refactor: move abi encoding code to noirc_abi
