+++ crates/nargo/src/errors.rs
@@ -1,5 +1,5 @@
 use hex::FromHexError;
-use noirc_abi::{errors::InputParserError, input_parser::InputValue, AbiType};
+use noirc_abi::errors::{AbiError, InputParserError};
 use std::{fmt::Display, io::Write, path::PathBuf};
 use termcolor::{Color, ColorChoice, ColorSpec, StandardStream, WriteColor};
 
@@ -52,40 +52,5 @@ impl From<InputParserError> for CliError {
 impl From<AbiError> for CliError {
     fn from(error: AbiError) -> Self {
         CliError::Generic(error.to_string())
-    }
-}
-
-#[derive(Debug)]
-pub enum AbiError {
-    Generic(String),
-    UnexpectedParams(Vec<String>),
-    TypeMismatch { param_name: String, param_type: AbiType, value: InputValue },
-    MissingParam(String),
-    UndefinedInput(String),
-}
-
-impl Display for AbiError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(
-            f,
-            "{}",
-            match self {
-                AbiError::Generic(msg) => msg.clone(),
-                AbiError::UnexpectedParams(unexpected_params) =>
-                    format!("Received parameters not expected by ABI: {:?}", unexpected_params),
-                AbiError::TypeMismatch { param_name, param_type, value } => {
-                    format!(
-                            "The parameter {} is expected to be a {:?} but found incompatible value {:?}",
-                            param_name, param_type, value
-                        )
-                }
-                AbiError::MissingParam(name) => {
-                    format!("ABI expects the parameter `{}`, but this was not found", name)
-                }
-                AbiError::UndefinedInput(name) => {
-                    format!("Input value `{}` is not defined", name)
-                }
-            }
-        )
     }
 }
