+++ crates/noirc_abi/src/encoding/mod.rs
@@ -1,0 +1,7 @@
+mod prover;
+mod verifier;
+
+pub use prover::process_abi_with_input;
+pub use verifier::process_abi_with_verifier_input;
+
+pub const MAIN_RETURN_NAME: &str = "return";
