+++ crates/noirc_abi/src/errors.rs
@@ -1,3 +1,4 @@
+use crate::{input_parser::InputValue, AbiType};
 use std::path::PathBuf;
 
 #[derive(Debug)]
@@ -20,5 +21,40 @@ impl std::fmt::Display for InputParserError {
             InputParserError::ParseHexStr(err_msg) => write!(f, "Could not parse hex value {}", err_msg),
             InputParserError::DuplicateVariableName(err_msg) => write!(f, "duplicate variable name {}", err_msg)
         }
+    }
+}
+
+#[derive(Debug)]
+pub enum AbiError {
+    Generic(String),
+    UnexpectedParams(Vec<String>),
+    TypeMismatch { param_name: String, param_type: AbiType, value: InputValue },
+    MissingParam(String),
+    UndefinedInput(String),
+}
+
+impl std::fmt::Display for AbiError {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(
+            f,
+            "{}",
+            match self {
+                AbiError::Generic(msg) => msg.clone(),
+                AbiError::UnexpectedParams(unexpected_params) =>
+                    format!("Received parameters not expected by ABI: {:?}", unexpected_params),
+                AbiError::TypeMismatch { param_name, param_type, value } => {
+                    format!(
+                            "The parameter {} is expected to be a {:?} but found incompatible value {:?}",
+                            param_name, param_type, value
+                        )
+                }
+                AbiError::MissingParam(name) => {
+                    format!("ABI expects the parameter `{}`, but this was not found", name)
+                }
+                AbiError::UndefinedInput(name) => {
+                    format!("Input value `{}` is not defined", name)
+                }
+            }
+        )
     }
 }
