+++ crates/noirc_frontend/src/hir_def/function.rs
@@ -13,8 +13,6 @@ use crate::{token::Attribute, FunctionKind};
 #[derive(Debug, Clone)]
 pub struct HirFunction(ExprId);
 
-pub const MAIN_RETURN_NAME: &str = "return";
-
 impl HirFunction {
     pub fn empty() -> HirFunction {
         HirFunction(ExprId::empty_block_id())
@@ -150,7 +148,7 @@ impl FuncMeta {
 
         if return_type != Type::Unit {
             let typ = return_type.as_abi_type(self.return_visibility);
-            abi.parameters.push((MAIN_RETURN_NAME.into(), typ));
+            abi.parameters.push((noirc_abi::encoding::MAIN_RETURN_NAME.into(), typ));
         }
 
         abi
